[********] fix(operator) 数组越界与不安全指针使用、依赖外部输入计算指针偏移
# message auto-generated for no-merge-commit merge:
merge 202503/FixIssue into master

安全问题修改，适配Mac clang编译

Created-by: insightDeveloper

Match-id-430aa470fa0f1455ddd51a2b62d5780aeae8a027

+++ server/src/modules/operator/handler/QueryOpStatisticInfoHandler.cpp
@@ -265,7 +265,7 @@ namespace Dic::Module::Operator {
         });
         uint64_t MAX_UINT64 = std::numeric_limits<uint64_t>::max();
         uint64_t safeTotal = total < 0 ? MAX_UINT64 : static_cast<uint64_t>(total);
-        safeTotal = std::min(safeTotal, statisticData.size());
+        safeTotal = std::min(safeTotal, static_cast<uint64_t>(statisticData.size()));
         std::vector<Protocol::OperatorStatisticCmpInfoRes> topKStatisticData(statisticData.begin(),
                                                                              statisticData.begin() + safeTotal);
